bioweb/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.1
django-bootstrap4==23.1
djangorestframework==3.14.0
factory-boy==3.0.1
Faker==18.10.1
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
